// ImpactaMais/src/app/models/ong-card.model.ts
export interface OngCard {
    id: number;
    nomeOng: string;
    causa: string;
    verificada: boolean;
    imagemCapa: string;
    avatarUrl: string;
    destaque: boolean;
    voluntariosCount: string;
    doacaoPercent: string;
    tituloCampanha: string;
    descricao: string;
    categoriaTag: string;
    localizacao: string;
}